feat: 럭키보드 공용 로직 공용 페이지로  이관 및 중복 api 호출 수정 (#121)
